pkg/endpoint: move identity / labels-related code
Move such code into identity.go and labels.go accordingly so such code is more
logically located within the endpoint package.

Signed-off by: Ian Vernon <ian@cilium.io>
